RMT: blocking transmit methods take &mut channel
Instead of taking the channel by value. This is much more convenient to
use, since it doesn't require unwrapping tuple errors, and doesn't
require re-assigning the channel that was previously consumed.

This is made safe with regard to `Drop`ping the transaction structs by
stopping and waiting in the Drop handler.

It's not perfect in case the transactions are leaked. It's possible to
make that case safe as well by extending the global STATE machine for
RMT channels. I'll leave that for another commit/PR.