RMT: add poll() method to blocking transactions
This allows properly interleaving several such transactions (as the HIL
loopback tests do, in fact) or other work.

Regarding the implementation of the place_rmt_driver_in_ram feature for these
methods, note that `#[ram]` implies `[inline(never)]`.

Thus, this uses `#[inline(always)]` for `poll` and `wait` in that case,
which are rather simple methods.
If putting them in ram actually matters, the calling user code should
probably be placed in ram as well, and then, forced inlining of both
methods should have the desired effect.

`poll_internal` in turn
- contains more code and is called from `poll` and `wait`
	-> it seems sensible that it is not inlined to reduce code size.
- includes the hot loop that copies to the hardware buffer
	-> should be forced to ram
so `#[ram]` makes sense.

Note that none of this was benchmarked.