RMT: process all pending IRQs in handle_channel_interrupts
Handle events for all channels in each invocation of the async_interrupt_handler.
FIXME: Is this actually a good idea? Think more about this and document
reasoning in a comment. Looping helps with fairness between channels,
but it might prevent other interrupts from being handled for quite some
time if several channels become ready simultaneously. On the other hand,
there are interrupt priorities to deal with this, and looping would
remove of multiple ISR entries in that case.