RMT: Add rmt_pin_reconfigure test that compiles, but shouldn't...
This code can be written because ChannelCreator's don't carry the RMT lifetime,
thus making it possible to move ChannelCreator out of Rmt, drop Rmt, and create
a new Rmt driver.

This became safe by accident when implementing the extended RAM feature
with its checks for Reserved memory blocks. However, it would be better
to prevent this at compile time. Additionally, the Reserved memory check
currently relies on mutable global static state, and we'd like to
refactor the code to get rid of that if possible.

This will be fixed in the next commit by properly tying together
lifetimes of RMT driver structs.