Improve async wifi performance
Before going into low power sleep, yield to the wifi task.
This should cause it to send outbound packets before sleeping.
It should also cause inbound packets to be processed earlier
since if we wake from waiti because of a wifi interrupt we will
repoll without progress and then run the wifi stack which will
generate work for the executor.