RMT: elide bounds checks via ChannelIndex enum
This removes (at least) the checks on

- indexing the STATE global
- register access via dynamic channel indices, which would otherwise
  contain a bounds check (implemented in the PAC)

by using a custom type for channel indices that can only take values for
which a channel exists.
Specifically, this uses an enum to serve as a refinement of u8 to help
the compiler restrict value ranges. Due to inlining, this restricted
value range is propagated by the compiler until the potential bounds checks,
even if there's intermediate cast to u8.
cf. https://github.com/rust-lang/rust/issues/109958