RMT: more consistent code formatting in low-level methods
Note that it's important how the closures for PAC register `modify` and
`write` calls are layed out. Compare

```rust
pub fn clear_tx_interrupts(&self) {
    let rmt = crate::peripherals::RMT::regs();

    rmt.int_clr().write(|w| {
	let ch_idx = self.ch_idx as u8;

	w.ch_tx_end(ch_idx).set_bit();
	w.ch_tx_err(ch_idx).set_bit();
	w.ch_tx_loop(ch_idx).set_bit();
	w.ch_tx_thr_event(ch_idx).set_bit()
    });
}
```

and

```rust
pub fn clear_tx_interrupts(&self) {
    let rmt = crate::peripherals::RMT::regs();
    let ch_idx = self.ch_idx as u8;

    rmt.int_clr().write(|w| {
	w.ch_tx_end(ch_idx).set_bit();
	w.ch_tx_err(ch_idx).set_bit();
	w.ch_tx_loop(ch_idx).set_bit();
	w.ch_tx_thr_event(ch_idx).set_bit()
    });
}
```

In the first example, the cast from `self.ch_idx: ChannelIndex` to `u8`
happens in the closure, whereas in the second case, the closure will be
built without the information about the restricted range of
`ChannelIndex` and assume `ch_idx` to be any `u8`. That leads to bounds
check being present (from the PAC register writer), which greatly
increases the closure size, and might push it above the compiler's
threshold for inlining. However, we generally rely quite a bit on
inlining with these methods to generate efficient code here and in
similar methods.