RMT: fix Channel/ChannelCreator drop
The code did properly track when to drop the GenericPeripheralGuard,
however, it did make the invalid assumption that it needs to release
channel memory under the same condition. That is incorrect when
ChannelCreator can be reborrowed, leading to not always releasing the
memory on Channel Drop when it should be.

Fixed here by explicitly tracking state with the DropState enum instead
of an Option, which was not sufficiently fine-grained.