RMT: in HIL tests, replace embassy by esp_rtos, reduce boilerplate
by factoring out more initialization code

Note that the refactoring of the rmt_use_all_channels tests makes it
fail on esp32 and esp32s2 due to reborrowing the ChannelCreator instead
of recreating the Rmt each time. This surfaces a bug that existed
before; subsequent commits will make sure this is better tested and add
a fix. The bug is not chip-specific, but doesn't show up on other
devices since the test uses different combinations of channels (without
reuse) for them.