RMT: lifetimes, tests, fixes (#4174)
* RMT: Add rmt_pin_reconfigure test that compiles, but shouldn't...

This code can be written because ChannelCreator's don't carry the RMT lifetime,
thus making it possible to move ChannelCreator out of Rmt, drop Rmt, and create
a new Rmt driver.

This became safe by accident when implementing the extended RAM feature
with its checks for Reserved memory blocks. However, it would be better
to prevent this at compile time. Additionally, the Reserved memory check
currently relies on mutable global static state, and we'd like to
refactor the code to get rid of that if possible.

This will be fixed in the next commit by properly tying together
lifetimes of RMT driver structs.

* RMT: track peripheral lifetimes properly, impl reborrow

- Rmt, ChannelCreator, Channel carry a lifetime: they're not necesarrily
  'static but might be reborrowed for a shorter lifetime
- impl ChannelCreator::reborrow and Channel::reborrow, which should
  unblock many RMT use cases by allowing re-configuration and by
  allowing rx/tx without consuming the Channel (i.e. a somewhat
  unergonomic variant of an API that would take the Channels by mutable
  reference)
- convert the rmt_pin_reconfigure test that previously compiled but
  shouldn't into a compile-only test that reborrowing ChannelCreator
  works

* RMT: avoid enable/disable churn on reborrow(), fix Channel::drop

Previously,

- Channel::drop unconditionally cleared the memory reservation for the
  channel, which became incorrect when introducing reborrowing
- each Channel::reborrow() or ChannelCreator::reborrow() would reference
  (and unref on drop) the Peripheral which introduces unnecessary
  overhead.

By only storing a GenericPeripheralGuard in the initial instance and not
any reborrowed version, this fixes both issues.

As a side effect, Rmt and ChannelCreator or not ZSTs anymore, but that
should be negligible.

* RMT: impl Drop for RxTransaction

* RMT: impl Drop for SingleShotTxTransaction

* RMT: impl Drop for ContinuousTxTransaction

* RMT: changelog + migration guide for Drop impls and #[must_use]

* RMT: RmtTxFuture and RmtRxFuture hold the channel lifetime

* RMT: rework rx/tx futures

- return the future directly instead of wrapping in a generated future
  from the async fn
- store errors and return them only on first poll: this helps avoid the
  somewhat inconvenient double unwrap/? to handle errors. Previously,
  this would automatically happens as part of the generated future.
- RmtRxFuture and RmtTxFuture capture the channel lifetime.
- require the data to be Unpin (which is unlikely to be a significant
  change, until very recently, the only accepted type was u32 anyway)

* RMT: impl Drop for RmtTxFuture

* RMT: impl Drop for RmtRxFuture

* RMT: changelog for impl Drop on futures

* RMT: add comments about generated tx data

* RMT: slightly more informative assertion in HIL tests

* RMT: add rmt_loopback_after_drop HIL test

* RMT: add rmt_loopback_after_drop_async HIL test

* RMT: add rmt_loopback_repeat_blocking HIL test

* RMT: add rmt_loopback_repeat_async HIL test

* RMT: re-enable previously broken esp32 HIL test

...to see whether it happens to be fixed

* RMT: stop_rx workaround for esp32, esp32s2

Fixes newly-added HIL tests that attempt new transactions after aborting
a previous one.

If we don't stop the receiver on Drop, subsequent receive() operations
will continue with partially old data. Unfortunately, esp32 and esp32s2
have no flag to forcefully stop the receiver, so we can only work around
this by trying to introduce a regular end condition.

This problem wasn't exposed before because the channel could only be
re-used in safe code after regularly ending a transaction and getting
back the channel via wait(). However, with reborrow() implemented, this
is now possible.

improve stop_rx

* RMT: revert to only a single lifetime for configure_tx/configure_rx

* RMT: in HIL tests, replace embassy by esp_rtos, reduce boilerplate

by factoring out more initialization code

Note that the refactoring of the rmt_use_all_channels tests makes it
fail on esp32 and esp32s2 due to reborrowing the ChannelCreator instead
of recreating the Rmt each time. This surfaces a bug that existed
before; subsequent commits will make sure this is better tested and add
a fix. The bug is not chip-specific, but doesn't show up on other
devices since the test uses different combinations of channels (without
reuse) for them.

* RMT: extend HIL repeat tests to cover more scenarios

* RMT: fix Channel/ChannelCreator drop

The code did properly track when to drop the GenericPeripheralGuard,
however, it did make the invalid assumption that it needs to release
channel memory under the same condition. That is incorrect when
ChannelCreator can be reborrowed, leading to not always releasing the
memory on Channel Drop when it should be.

Fixed here by explicitly tracking state with the DropState enum instead
of an Option, which was not sufficiently fine-grained.

* RMT: share more code in HIL tests

by factoring out running + checking the loopback transaction everywhere

---------

Co-authored-by: Juraj Sadel <juraj.sadel@espressif.com>