RMT: Merge interrupt flag reset methods into clear_?x_interrupts
By adding an EnumSet argument. This nicely mirrors the listen/unlisten
functions, which take the same type of argument. It also helps to avoid
proliferation of a large number of methods when using more interrupts
(such as the loopcount interrupt, which I intend to do in a later PR).

There should be no performance negative, since these methods are always
inlined with compile-time known arguments, such that the compiler can
remove all conditionals.