RMT: change method receivers for DynChannelAccess from &self to self
Since it is Copy. We expect all of this to be inlined, so there
shouldn't be a difference in the final code. This should however
simplify the compiler's job a tiny bit and more importantly prevent
pointer indirection and spilling to stack if one of these methods ends
up being not inlined for whatever reason.