RMT: rewrite clear_*x_interrupt in obviously branch-free manner
This make the code slightly more concise and obviously branch-free
(although the compiler seems to make that optimization on it's own, as
well).

This code remains equivalent, because int_clr has only a write-to-clear
bits and the register writer will be initialized to all zeroes before
the closure is invoked.