RMT: only store &mut dyn EncoderExt in transaction structs
Previously, we stored &mut dyn Encoder and dynamically dispatched the Encoder::encode method.
Now, we store &mut dyn EncoderExt and dynamically dispatch
EncoderExt::write with the expectation that Encoder::encode should be
inlined in EncoderExt::write (which is the only caller, and its Encoder
implementations in esp-hal are also marked as #[inline(always)]).

This might allow for small optimizations, since the RmtWriter type will
typically not need to be constructed on the stack, but can be kept in
registers.