RMT: add HIL tests for more tx error cases
to ensure that upcoming Iterator/Encoder-related refactoring dont't break anything.

The EndMarkerMissing cases with different tx lengths seem redundant
right now (since this is verified before even starting the
transmission), but that will change when iterator input is supported and
the presence of an end marker can only be verified when the iterator is
exhausted.