RMT: test cleanups (#4493)
* RMT: slight change to and better explanation for generated tx data

* RMT: remove rmt_single_shot_* HIL tests

I added them at some point for debugging, but they have little value in
addition to the loopback tests which are being run for all the same
settings (but can actually verify that tx works)

* RMT: use esp-alloc in HIL tests to avoid generics

* RMT: in HIL tests, explicitly base data size on CHANNEL_RAM_SIZE

* RMT: refactor HIL tests to simplify function signatures

Helper methods had grown quite a few arguments that aren't easy to
disambiguate just by their type, making calls of those functions hard to
read.
Thus, introduce a config struct to pass around these settings in
aggregate and refer to them by their field name.

* RMT: factor out BitbangTransmitter out of rmt_clock_rate HIL test

There was a bit too much going on in this test to be able to understand
it easily.

* RMT: add support for tests with end marker in the second PulseCode field

* RMT: add a HIL test with the end marker in field 2

* RMT: Remove outdated restriction from docs

* RMT: update explanatory comments for loopback tests