RMT: Refactor Drop implementations with guard types (#4492)
* RMT: Channel drop via MemoryGuard

instead of implementing Drop on Channel and using the DropState enum to
decide which cleanup actions are required, place the
GenericPeripheralGuard and a new MemoryGuard in the Channel struct which
are concerned with only one part of the cleanup

* RMT: dedup transaction Drop impls via guard

The various (Blocking/Async, one-shot/continous) transaction all had
identical Drop implementations. This shares this code by moving it into
guard types. These guards can be disabled to avoid unnecessary cleanup
actions if the transactions are known to have completed already or were
never started due to an error. This also removes the use of ManuallyDrop
and unsafe code to access the ManuallyDrop content.

This also replaces the clear_*x_interrupts call on Drop with
unlisten_*x_interrupt: This seems crucial in case of dropping async
futures. (For example consider the case of dropping a future, then
starting a blocking transaction: while probably a bit of an edge case, previously
the interrupts would still have been enableed.)
clear_*x_interrupts is called in any case before starting a new transaction.

This also removes an unnecessary update call and adds a missing
stop_rx() to the RxFuture poll() method. (This was previously handled by
RxFuture::drop(), but a bit poorly understood how things interacted. Also added
some comments to explain things a bit better.)