MVP for light/deep sleep with timer wakeup
The design is based on C6 implementation of esp-hal,
and the logic is based on H2 implementation of esp-idf.
It is worth noting that esp-idf implementation writes
to a lot of registers which are not documented in
H2 Technical Reference Manual -- these writes are not
added here, which does not break functionality.

Both light sleep and deep sleep seem to work in an
example app with BT connection, with wakeup on timer.
No other wakeup source is currently supported.

Some other limitations:
  * clock settings are not preserved through light sleep
  * peripheral clocks are not disabled in light sleep