RMT: revise error handling, track total in writer, make writer.state read-only
- shares a tiny bit more code between Blocking and Async tx
- removes Slice-specific error handling in preparation for supporting
  iterator (-like) data types
- makes RmtWriter.state read-only and only write it from the rmt::write
  module to make it easier to reason about

As a side effect, continuous tx will now also report EndMarkerMissing
errors. This makes sense: The loopcount interrupt only appears to
trigger when hitting and end marker, and does not trigger when in
loop tx mode with wrapping enabled and wrapping at the buffer end. In
other words, in that case, loop tx seems to work at first, but does not
trigger interrupts, so it should probably be prevented (it might make
sense to allow LoopMode::Infinite without end marker, however).